refactor: 레이어별 프로젝트 구조 세팅
